requirements.txt
mcp-agent
anthropic
PyGithub
GitPython
pytest
